getdents on FreeBSD returns ssize_t

+++ lib/libc/sys/getdents.c
@@ -1,5 +1,5 @@
 #include <dirent.h>
 
-int getdents(int fd, char *buf, size_t nbytes) {
+ssize_t getdents(int fd, char *buf, size_t nbytes) {
   return getdirentries(fd, buf, nbytes, NULL);
 }
